bugfix: fix module order to avoid taking effect before other modules such as sub filter module (#31)
* bugfix: add module order to avoid taking effect before other modules such as sub filter module

Module order is incorrect for compression module. For example, if using ngx_http_sub_module, string is not replaced because it it tries to replace content when it is already compressed.

For this purpose, you need to add module sorting in config to adjust the order in which modules take effect.